Create BreakingTheRecords.java
Maria plays college basketball and wants to go pro. Each season she maintains a record of her play. She tabulates the number of times she breaks her season record for most points and least points in a game. Points scored in the first game establish her record for the season, and she begins counting from there.

Returns a List<Integer> of length 2. The first value is the amount of times Maria broke her record for highest score. The second value is the amount of times that Maria broke her record for lowest score.